+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+
+### Deleted
+
+* disabled overrides
+
+### Changed
+
+* project types destroy action to be consistent with Redmine 4.2
+* structure in lib dir
+
 ## 4.0.3 - 2021-04-22
 
 ### Added